api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pillow>=8.0.0
scikit-image>=0.19.0
numpy>=1.20.0
pydantic>=2.0.0
fastmcp>=0.0.9
matplotlib>=3.5.0
python-jose>=3.3.0
httpx>=0.24.0